// src/app/services/models/scheduled-transfer-request-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ScheduledTransferRequestDto {
  amount: number;
  description?: string;
  receiverEmail: string;
  recurrenceEndDate?: string;
  recurrenceType?: 'NONE' | 'DAILY' | 'WEEKLY' | 'MONTHLY';
  scheduledDateTime: string;
  senderEmail: string;
  totalOccurrences?: number;
}