// src/app/services/models/user-representation.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CredentialRepresentation } from '../models/credential-representation';
import { FederatedIdentityRepresentation } from '../models/federated-identity-representation';
import { SocialLinkRepresentation } from '../models/social-link-representation';
import { UserConsentRepresentation } from '../models/user-consent-representation';
import { UserProfileMetadata } from '../models/user-profile-metadata';
export interface UserRepresentation {
  access?: {
[key: string]: boolean;
};
  /** @deprecated */applicationRoles?: {
[key: string]: Array<string>;
};
  attributes?: {
[key: string]: Array<string>;
};
  clientConsents?: Array<UserConsentRepresentation>;
  clientRoles?: {
[key: string]: Array<string>;
};
  createdTimestamp?: number;
  credentials?: Array<CredentialRepresentation>;
  disableableCredentialTypes?: Array<string>;
  email?: string;
  emailVerified?: boolean;
  enabled?: boolean;
  federatedIdentities?: Array<FederatedIdentityRepresentation>;
  federationLink?: string;
  firstName?: string;
  groups?: Array<string>;
  id?: string;
  lastName?: string;
  notBefore?: number;
  origin?: string;
  realmRoles?: Array<string>;
  requiredActions?: Array<string>;
  self?: string;
  serviceAccountClientId?: string;
  /** @deprecated */socialLinks?: Array<SocialLinkRepresentation>;
  /** @deprecated */totp?: boolean;
  userProfileMetadata?: UserProfileMetadata;
  username?: string;
}